+++ vite.config.ts
@@ -9,7 +9,7 @@ export default defineConfig({
       formats: ['es'],
     },
     rollupOptions: {
-      // Cesium 是对等依赖，库包不能把它打进去。
+      // Cesium 是对等依赖，库包不能把它打进去
       external: ['cesium'],
     },
     sourcemap: true,